Back to 12 bit value for angle to avoid cofusion with software which was not adjusted. Switching angleD to float instead of short to avoid catastrofic loss of performance while averaging for fast update rate.

+++ Firmware/CartPole CubeIDE/Src/angle_processing.h
@@ -8,8 +8,8 @@
 #ifndef ANGLE_PROCESSING_H_
 #define ANGLE_PROCESSING_H_
 
-void average_derivatives(int* angleDPtr, short* positionDPtr);
-void process_angle(int angleSamples[], unsigned short angleSampIndex, unsigned short angle_averageLen, int* anglePtr, int* angleDPtr, int* invalid_stepPtr);
+void average_derivatives(float* angleDPtr, short* positionDPtr);
+void process_angle(int angleSamples[], unsigned short angleSampIndex, unsigned short angle_averageLen, int* anglePtr, float* angleDPtr, int* invalid_stepPtr);
 int anomaly_detection(int* angleSamples, unsigned short angleSampIndex, unsigned short angle_averageLen);
 void treat_deadangle_with_derivative(int* anglePtr, int invalid_step);
